modificacion de la api método update de lecturas y actualizar código arduino para enviar los post a la api
No era necesario el método updale Lecturas, lo eliminé para que el servicio cree el GUID automáticamente y no lo haga el micro controlador . el micro escribe en la base de datos correctamente las lecturas.

+++ Código/Web/SiloBolsa.Core/Modelos/Lectura.cs
@@ -14,7 +14,7 @@ namespace SiloBolsa.Core.Modelos
         public DateTime FechaHoraLectura { get; set; }
 
         [Column("temp")]
-        public int Temp { get; set; }
+        public double Temp { get; set; }
 
         [Column("humedad")]
         public double Humedad { get; set; }
